# Improves instruction flattener algorithm