On my low-powered laptop, this brought the load time of firmware.wasm from 1090 msec to 896 msec (82%), of which the flattener went from 653 msec to 396 msec (61%).

See benchmark results for the Flattener in the sub-commits.